BN-68 Separated a part of addon to use it in CI and prepared a simple test
